add filter to exporter queryset
