+++ django_project/dashboard/tools/entity_query.py
@@ -1,7 +1,6 @@
 from typing import Tuple, List
-from dateutil.parser import parse
-from datetime import datetime
 from django.contrib.gis.geos import Point
+from georepo.utils.entity_query import validate_datetime
 from georepo.models import (
     Dataset, DatasetView
 )
@@ -9,17 +8,6 @@ from georepo.models import (
 from dashboard.models import (
     EntitiesUserConfig
 )
-
-
-def validate_datetime(value: str) -> datetime:
-    res = None
-    if value is None:
-        return res
-    try:
-        res = parse(value)
-    except ValueError:
-        pass
-    return res
 
 
 def generate_query_condition(
@@ -86,43 +74,28 @@ def generate_query_condition(
         sql = (
             sql + 'AND gg.privacy_level IN %s ')
         query_values.append(tuple(filter.filters['privacy_level']))
+    if 'source' in filter.filters and \
+        len(filter.filters['source']) > 0:
+        sql = (
+            sql + 'AND gg.source IN %s ')
+        query_values.append(tuple(filter.filters['source']))
     if ('search_text' in filter.filters and
             len(filter.filters['search_text']) > 0):
         sql = (
             sql + 'AND (gg.label ilike %s OR '
-            'ge.label ilike %s OR '
             'parent_0.label ilike %s OR '
             'gg.unique_code ilike %s OR '
             'gg.concept_ucode ilike %s OR '
             'ge_id.value ilike %s OR '
             'ge_name.name ilike %s  OR '
-            'gg.unique_code ilike %s  OR '
             'gg.internal_code ilike %s  OR '
-            'gg.label ilike %s  OR '
             'gg.source ilike %s  OR '
-            'gg.admin_level_name ilike %s '
+            'gg.admin_level_name ilike %s)'
         )
 
         search_text = '%' + filter.filters['search_text'] + '%'
-        for i in range(12):
+        for i in range(9):
             query_values.append(search_text)
-
-        try:
-            search_text_int = int(filter.filters['search_text'])
-            sql = (
-                sql + 'OR gg.revision_number = %s OR '
-                'gg.version = %s OR '
-                'gg.feature_units = %s OR '
-                'gg.vertices = %s '
-                ')'
-            )
-            for i in range(4):
-                query_values.append(search_text_int)
-        except ValueError:
-            sql = (
-                sql + ')'
-            )
-            pass
 
     if 'points' in filter.filters:
         points_cond = []
